# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /knowledge/search:
    post:
      operationId: searchKnowledge
      tags: [knowledge]
      summary: Diagnostic lexical knowledge search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                purpose:
                  type: string
                travelerType:
                  type: string
                limit:
                  type: integer
              required: [query]
      responses:
        "200":
          description: Search hits with scores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeSearchResponse"
  /knowledge/import:
    post:
      operationId: importKnowledgeWorkbook
      tags: [knowledge]
      summary: Import a SIT knowledge workbook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fileName:
                  type: string
                dataBase64:
                  type: string
              required: [fileName, dataBase64]
      responses:
        "200":
          description: Import metadata and report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeImportResponse"
components:
  schemas:
...
    KnowledgeImportMetadata:
      type: object
      properties:
        version:
          type: string
        importedAt:
          type: string
        sourceWorkbook:
          type: string
        sheetName:
          type: string
        cardCount:
          type: integer
      required: [version, importedAt, sourceWorkbook, sheetName, cardCount]
    KnowledgeCard:
      type: object
      properties:
        id: { type: string }
        category: { type: string }
        subcategory: { type: string }
        topic: { type: string }
        travelerType: { type: string }
        myth: { type: string }
        reality: { type: string }
        deepLocalInsight: { type: string }
        psychologicalLayer: { type: string }
        recommendedFor: { type: string }
        notRecommendedFor: { type: string }
        aiRecommendationLogic: { type: string }
        relatedCategories:
          type: array
          items: { type: string }
        hiddenBenefit: { type: string }
        hiddenRisk: { type: string }
        updateFrequency: { type: string }
        confidence: { type: string }
        priorityScore: { type: number }
        sourceExpert: { type: string }
        architectureLayer: { type: string }
        masterCategory: { type: string }
        dataMoatScore: { type: number }
        monetizationScore: { type: number }
        investorValueScore: { type: number }
        localityScore: { type: number }
        status: { type: string }
        version: { type: string }
        importedAt: { type: string }
      required: [id, category, topic, version, importedAt]
...
    KnowledgeSearchResponse:
      type: object
      properties:
        version:
          type: string
        importedAt:
          type: string
        hits:
          type: array
          items:
            type: object
            properties:
              card:
                $ref: "#/components/schemas/KnowledgeCard"
              score:
                type: number
      required: [version, hits]
    KnowledgeImportResponse:
      type: object
      properties:
        metadata:
          $ref: "#/components/schemas/KnowledgeImportMetadata"
        report:
          type: object
        files:
          type: object
      required: [metadata, report]